console: Move containers folder to console
