+++ pkg/webui/console/containers/applications-table/applications-table.styl
