+++ pkg/webui/console/containers/gateways-table/index.js
@@ -12,65 +12,72 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { Component } from 'react'
+import React from 'react'
 import { defineMessages } from 'react-intl'
 import bind from 'autobind-decorator'
 
+import sharedMessages from '../../../lib/shared-messages'
 import FetchTable from '../fetch-table'
 
 import {
-  getApplicationsList,
-  searchApplicationsList,
-} from '../../console/store/actions/applications'
+  getGatewaysList,
+  searchGatewaysList,
+} from '../../../console/store/actions/gateways'
 
 const m = defineMessages({
-  all: 'All',
-  appId: 'Application ID',
-  desc: 'Description',
-  empty: 'No items matched your criteria',
-  add: 'Add Application',
+  add: 'Add Gateway',
+  gtwId: 'Gateway ID',
+  gtwEUI: 'Gateway EUI',
+  freqPlan: 'Frequency Plan',
 })
-
-const tabs = [
-  {
-    title: m.all,
-    name: 'all',
-    disabled: true,
-  },
-]
 
 const headers = [
   {
-    name: 'ids.application_id',
-    displayName: m.appId,
+    name: 'ids.gateway_id',
+    displayName: m.gtwId,
+    width: 45,
   },
   {
-    name: 'description',
-    displayName: m.desc,
+    name: 'ids.eui',
+    displayName: m.gtwEUI,
+    width: 25,
+  },
+  {
+    name: 'name',
+    displayName: sharedMessages.name,
+    width: 10,
+  },
+  {
+    name: 'antennasCount',
+    displayName: sharedMessages.antennas,
+    centered: true,
+    width: 5,
+  },
+  {
+    name: 'frequency_plan_id',
+    displayName: m.freqPlan,
+    width: 15,
   },
 ]
 
 @bind
-export default class ApplicationsTable extends Component {
+export default class GatewaysTable extends React.Component {
 
-  baseDataSelector ({ applications }) {
-    return applications
+  baseDataSelector ({ gateways }) {
+    return gateways
   }
 
   render () {
     return (
       <FetchTable
-        entity="applications"
-        headers={headers}
+        entity="gateways"
         addMessage={m.add}
-        tableTitle={this.tableTitle}
-        getItemsAction={getApplicationsList}
-        searchItemsAction={searchApplicationsList}
-        tabs={tabs}
+        headers={headers}
+        getItemsAction={getGatewaysList}
+        searchItemsAction={searchGatewaysList}
         baseDataSelector={this.baseDataSelector}
         {...this.props}
       />
     )
   }
 }
-
